+++ CountlyConnectionManager.h
@@ -35,7 +35,7 @@ extern NSString* const kCountlyQSKeyDeviceID;
 - (void)sendCrashReport:(NSString *)report immediately:(BOOL)immediately;
 - (void)sendOldDeviceID:(NSString *)oldDeviceID;
 - (void)sendParentDeviceID:(NSString *)parentDeviceID;
-- (void)sendAttribution:(NSString *)attribution;
+- (void)sendAttribution;
 - (void)sendConsentChanges:(NSString *)consentChanges;
 - (void)sendPerformanceMonitoringTrace:(NSString *)trace;